[MoveOnly] Move non-wallet RPC files to subdir
No code changes here, just moving of non-wallet rpc files to their own
subdir.
