+++ src/rpc/blockchain.cpp
@@ -9,7 +9,7 @@
 #include "checkpoints.h"
 #include "clientversion.h"
 #include "main.h"
-#include "rpcserver.h"
+#include "rpc/server.h"
 #include "sync.h"
 #include "txdb.h"
 #include "util.h"
